Add State service
