+++ config/packages/twig.yaml
@@ -4,3 +4,5 @@ twig:
     paths:
         - '%kernel.project_dir%/templates'
     exception_controller: App\Controller\ExceptionController::showAction
+    globals:
+        state: '@App\Request\State'
